refactor: RootFeature 모듈을 MainTabFeature로 명확화
