+++ Projects/App/Sources/AppCore/Navigation/AppRootFlowView.swift
@@ -2,7 +2,7 @@ import AuthFeature
 import ComposableArchitecture
 import DSKit
 import OnboardingFeature
-import RootFeature
+import MainTabFeature
 import SwiftUI
 
 struct AppRootFlowView: View {
